// CreditDebit/src/types/schema.ts
// THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.

import { Document } from "@blockflow-labs/utils";

export class ERC20Table {
  static entity = "ERC20Table";
  static schema = {
    id: { type: "String", index: true },
    address: "string",
    counterPartyAddress: "string",
    tokenAddress: "string",
    tokenName: "string",
    tokenSymbol: "string",
    rawAmount: "number",
    rawAmountString: "string",
    amount: "string",
    amountString: "string",
    usdValue: "number",
    usdExchangeRate: "number",
    transactionHash: "string",
    logIndex: "number",
    blockTimestamp: "string",
    blockNumber: "number",
    blockHash: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

import { String, Array } from "@blockflow-labs/utils";

export interface IERC20Table extends Document {
  id: String;
  address: string;
  counterPartyAddress: string;
  tokenAddress: string;
  tokenName: string;
  tokenSymbol: string;
  rawAmount: number;
  rawAmountString: string;
  amount: string;
  amountString: string;
  usdValue: number;
  usdExchangeRate: number;
  transactionHash: string;
  logIndex: number;
  blockTimestamp: string;
  blockNumber: number;
  blockHash: string;
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}